+++ test/unit/plugins/communicators/winrm/helper_test.rb
@@ -38,12 +38,73 @@ describe VagrantPlugins::CommunicatorWinRM::Helper do
     end
   end
 
+  describe ".winrm_info" do
+    before do
+      machine.provider.stub(:capability?).
+        with(:winrm_info).and_return(false)
+      subject.stub(winrm_address: nil)
+      subject.stub(winrm_port: nil)
+    end
+
+    it "returns default info if no capability" do
+      subject.stub(winrm_address: "bar")
+      subject.stub(winrm_port: 45)
+
+      result = subject.winrm_info(machine)
+      expect(result[:host]).to eq("bar")
+      expect(result[:port]).to eq(45)
+    end
+
+    it "raises an exception if capability returns nil" do
+      machine.provider.stub(:capability?).
+        with(:winrm_info).and_return(true)
+      machine.provider.stub(:capability).with(:winrm_info).and_return(nil)
+
+      expect { subject.winrm_info(machine) }.
+        to raise_error(VagrantPlugins::CommunicatorWinRM::Errors::WinRMNotReady)
+    end
+
+    it "returns the proper information if set" do
+      machine.provider.stub(:capability?).
+        with(:winrm_info).and_return(true)
+      machine.provider.stub(:capability).with(:winrm_info).and_return({
+        host: "foo",
+        port: 12,
+      })
+
+      result = subject.winrm_info(machine)
+      expect(result[:host]).to eq("foo")
+      expect(result[:port]).to eq(12)
+    end
+
+    it "defaults information if capability doesn't set it" do
+      machine.provider.stub(:capability?).
+        with(:winrm_info).and_return(true)
+      machine.provider.stub(:capability).with(:winrm_info).and_return({})
+
+      subject.stub(winrm_address: "bar")
+      subject.stub(winrm_port: 45)
+
+      result = subject.winrm_info(machine)
+      expect(result[:host]).to eq("bar")
+      expect(result[:port]).to eq(45)
+    end
+  end
+
   describe ".winrm_port" do
     it "returns the configured port if no guest port set" do
       machine.config.winrm.port = 22
       machine.config.winrm.guest_port = nil
 
       expect(subject.winrm_port(machine)).to eq(22)
+    end
+
+    it "returns the configured guest port if non local" do
+      machine.config.winrm.port = 22
+      machine.config.winrm.guest_port = 2222
+      machine.config.vm.network "forwarded_port", host: 1234, guest: 2222
+
+      expect(subject.winrm_port(machine, false)).to eq(2222)
     end
 
     it "returns a forwarded port that matches the guest port" do